fastapi-ai-server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
langchain==0.0.184
openai==0.27.0
python-dotenv==0.19.0